+++ src/pm/util/pmiserv.c
@@ -1391,7 +1391,7 @@ static int PMIUBufferedReadLine( PMIProcess *pentry, char *buf, int maxlen )
                 /* Carefully read data into buffer.  This could be
                    written to read more at one time, but would then
                    need to know the size of the readbuf */
-		n = read( fd, readbuf, 1 );
+		n = (int)read( fd, readbuf, 1 );
 	    } while (n == -1 && errno == EINTR);
 	    if (n == 0) {
 		/* EOF */
